Add files via upload
This is the Car Class with revisions made to the setDistance(), setPath(), and setSpeed() methods. 

I have a question about the setPath() method, since we are using an ArrayDeque, we want to pop() the first item in the Deque to be the path, correct?  Do I need to cast that to a String?